registry: revert opam/k3kcli backends to ubi (#6406)
Resolves https://github.com/jdx/mise/discussions/6402.

Neither works with the GitHub backend because the releases have
archives, but we need to download non-archive files.
https://github.com/ocaml/opam/releases/tag/2.4.1
https://github.com/rancher/k3k/releases/tag/v0.3.4

We should fix it, but I don't have a good idea how to fix it.
Maybe we can only exclude small non-archive files (text files) when
archive files exist, but I think this also causes some bugs.