Flush vfox download buffer
I suspect (but can't confirm) the issue in
https://github.com/jdx/mise/discussions/6527 is due to a race condition
when writing the downloaded archive to disk.

If the write isn't fully flushed to disk, it's possible for readers to
see an incomplete/invalid archive. As a precaution, we should explicitly
flush before returning.

The issue is probably more likely on CI runners with slower file
systems.

As for why we seem to only see this with `yarn`, I'm guessing its
smaller size (1.2M) makes it more likely to sit in OS buffers longer.

This is all just a theory! Again I haven't been able to reliably
reproduce the issue. Adding a flush seems like it can't hurt, though.