docs: simplify apt instructions (#6986)
Container recipes run as root within the container, so there's no need
for sudo in them. In some other contexts sudo was already being used
e.g. with `sudo apt-get update`, so it was implicitly already installed.

The ASCII armored .pub can be used in keyrings as-is, no need for gpg
conversion, so no need to install gpg.

Use curl instead of wget everywhere, so there's no need to install wget.