fix(pwsh): wrap the executable path with double quotes (#6993)
It ocurred to me that the script generated by running `mise activate
pwsh` throws an error because my user home contains whitespace and the
generated script didn't escape them.

A portion of the current generated script:

```pwsh
    if ($arguments.count -eq 0) {
        & C:\Users\User Home\AppData\Local\Microsoft\WinGet\Links\mise.exe
        _reset_output_encoding
        return
    } elseif ($arguments -contains '-h' -or $arguments -contains '--help') {
        & C:\Users\User Home\AppData\Local\Microsoft\WinGet\Links\mise.exe @arguments
        _reset_output_encoding
        return
    }
```

Which cause the error:

```
&:
Line |
  64 |          & C:\Users\User Home\AppData\Local\Microsoft\WinGet\Lin …
     |            ~~~~~~~~~~~~~~
     | The term 'C:\Users\User' is not recognized as a name of a cmdlet, function, script file, or executable program.
Check the spelling of the name, or if a path was included, verify that the path is correct and try again.
```

To fix it the executable path should be wrapped with quotes in every
ocassion

```pwsh
    if ($arguments.count -eq 0) {
        & "C:\Users\User Home\AppData\Local\Microsoft\WinGet\Links\mise.exe"
        _reset_output_encoding
        return
    } elseif ($arguments -contains '-h' -or $arguments -contains '--help') {
        & "C:\Users\User Home\AppData\Local\Microsoft\WinGet\Links\mise.exe" @arguments
        _reset_output_encoding
        return
    }
```