fix: in `activate bash` output, wrap mise executable path in single-quotes (#7002)
On Windows, the path to `mise.exe` will contain backslashes.
Both `busybox.exe bash` and Git-for-Windows' "Git Bash" can understand
these backslash paths.
But bash syntax rules dictate that unquoted backslashes are interpreted
as escapes. We must wrap the path in single-quotes to avoid this.

```bash
# In "Git Bash" or "busybox.exe bash"
# Backslash paths are accepted
$ 'G:\dev\@jdx\mise\target\debug\mise.exe' --version
...
2025.11.5-DEBUG windows-x64 (2025-11-18)

# But unqouted is broken
# because bash's syntax rules will interpret any unquoted backslashes as escapes
$ G:\dev\@jdx\mise\target\debug\mise.exe --version
bash: G:dev@jdxmisetargetdebugmise.exe: not found
```

Quoting is *technically* more correct on Linux, too, even though the
issues it avoids -- splitting on spaces -- are largely hypothetical.[^1]

I can't think of any way that this code change would break someone's
legitimate setup on Linux or Mac.

[^1]: To be pedantic, if the path to `mise` includes a single-quote --
if the name of a parent directory has a single-quote in it -- then this
will be broken on Linux, but it would *already* be broken before this
code change, so we are not making it worse.