fix: silence kerl update
I work on https://railpack.com which heavily uses mise (we love it!). One of the things
we do is parse `mise latest` output. However, on the initial run of `mise latest erlang`
the `kerl update` command is run and the stdout is redirected to the process stdout, which
causes the command output to include the kerl update results. For example:

```
Getting releases from GitHub...
The available releases are:
17.5.6.10
18.3.4.11
19.3.6.13
20.3.8.26
21.3.8.24
22.3.4.27
23.3.4.20
24.3.4.17
25.3.2.21
26.2.5.16 *
27.3.4.6 *
28.3 *
28.3
```

This causes two problems:

1. `mise latest erlang*` generates indeterminent results. First run looks different than the second.
2. For our specific purposes, this makes it hard to parse.

This change redirects `kerl update` stdout + stderr to the trace logs.